+++ code_index/mcp_server/services/__init__.py
@@ -15,5 +15,5 @@ Classes:
 """
 
 from .code_index_service import CodeIndexService
-from .llm_traverse_codebase import RepoAnalyseService
+from .repo_analyse_service import RepoAnalyseService
 from .source_code_fetch_service import SourceCodeFetchService
